fix: progress for copying source dir
